Changes to make the FLAT docker setup work with native Docker for Mac and with the latest phusion/baseimage.
The SWORD interface is now configured for localhost by default (as that will work for the all Docker hosts: native Mac/Win and *nix). One can adapt this by changing the FLAT_HOST environment variable in the flat/Dockerfile.

phusion/baseimage is now based on Ubuntu 16.04 which only supports PHP 7 and provides Java 8 out of the box. FLAT is now based on a specific baseimage version (0.9.19) instead of the latest to prevent suprises.